Changed int to float

+++ automation_functions.py
@@ -26,7 +26,7 @@ def get_driver() -> webdriver:
     return webdriver.Chrome(path_to_script, options=chrome_options)
 
 
-def check_xpath_element(driver: webdriver, xpath: str, wait_time: int = 2) -> Optional[type(webdriver.remote)]:
+def check_xpath_element(driver: webdriver, xpath: str, wait_time: float = 2) -> Optional[type(webdriver.remote)]:
     try:
         element: webdriver.remote = WebDriverWait(driver, wait_time).until(
             EC.presence_of_element_located(("xpath", xpath)))
@@ -37,7 +37,7 @@ def check_xpath_element(driver: webdriver, xpath: str, wait_time: int = 2) -> Op
     return element
 
 
-def check_xpath_elements(driver: webdriver, xpath: str, wait_time: int = 2) -> Optional[type(webdriver.remote)]:
+def check_xpath_elements(driver: webdriver, xpath: str, wait_time: float = 2) -> Optional[type(webdriver.remote)]:
     try:
         elements: webdriver.remote = WebDriverWait(driver, wait_time).until(
             EC.presence_of_all_elements_located(("xpath", xpath)))
@@ -48,7 +48,7 @@ def check_xpath_elements(driver: webdriver, xpath: str, wait_time: int = 2) -> O
     return elements
 
 
-def scroll_down(driver: webdriver, scroll_value: int = 325) -> None:
+def scroll_down(driver: webdriver, scroll_value: float = 325) -> None:
     sleep(0.5)
     driver.execute_script("window.scrollTo(0, window.scrollY + {});".format(scroll_value))
     sleep(0.5)
